remove pyop wrapper
