+++ onnxruntime/core/language_interop_ops/pyop/pyop.h
@@ -19,6 +19,7 @@ using OnnxTypes   = std::vector<ONNXTensorElementDataType>;
 using OnnxAttrs   = std::unordered_map<std::string, std::string>;
 using PyOpLogFunc = std::function<void(const char*)>;
 
+/*
 typedef bool Initialize();
 typedef void ReleaseInstance(void*);
 typedef bool InvokePythonFunc(void*,
@@ -46,6 +47,30 @@ public:
 private:
     PyOpLibProxy();
     ~PyOpLibProxy();
+};
+*/
+
+class PyOpLibProxy {
+
+public:
+    static PyOpLibProxy& GetInstance();
+    void ReleaseInstance(void*);
+    bool InvokePythonFunc(void*,
+                          const char*,
+                          const std::vector<const void*>&,
+                          const std::vector<int32_t>&,
+                          const std::vector<std::vector<int64_t>>&,
+                          std::vector<std::unique_ptr<char[]>>&,
+                          std::vector<int32_t>&,
+                          std::vector<std::vector<int64_t>>&,
+                          std::function<void(const char*)>);
+    const char* GetLastErrorMessage(std::string&);
+    void* NewInstance(const char*, const char*, const OnnxAttrs&);
+    bool Initialized() const { return initialized_; }
+private:
+    PyOpLibProxy();
+    ~PyOpLibProxy();
+    bool initialized_ = false;
 };
 
 struct PyCustomKernel {
